react-web created

+++ services/profile-api/Services/Concrete/ProfileService.cs
@@ -18,9 +18,9 @@ public class ProfileService(
     HttpClient client) : IProfileService
 {
     private readonly IDatabase _cache = redis.GetDatabase();
-    private readonly string _mediaApiUrl = configuration.GetValue<string>("MediaApiUrl")!;
+    private readonly string _mediaApiUrl = configuration["MediaService:Url"]!;
 
-    public async Task<ServiceResponse<ProfileDto>> GetProfile(string username, IFormFileCollection files)
+    public async Task<ServiceResponse<ProfileDto>> GetProfile(string username)
     {
         var cacheKey = $"profile_{username}";
         var cachedProfile = await _cache.StringGetAsync(cacheKey);
@@ -44,19 +44,16 @@ public class ProfileService(
 
     public async Task<ServiceResponse<NoContent>> Update(ProfileDto model)
     {
+        var cacheKey = $"profile_{model.Id}";
+
         var profile = await context.Profiles.FirstOrDefaultAsync(x => x.Id == model.Id);
         if (profile == null) return ServiceResponse<NoContent>.Failure("Profile Not Found", StatusCodes.Status404NotFound);
         profile.Bio = model.Bio;
         profile.Username = model.Username;
         profile.FullName = model.FullName;
-        // profile.Email = model.Email;
-        //profile.AvatarPath = model.AvatarPath;
         context.Update(profile);
         await context.SaveChangesAsync();
-        
-        var cacheKey = $"profile_{model.Id}";
         await _cache.KeyDeleteAsync(cacheKey);
-        
         await _cache.StringSetAsync(cacheKey, JsonSerializer.Serialize(profile));
         return ServiceResponse<NoContent>.Success(StatusCodes.Status200OK);
     }
@@ -64,12 +61,12 @@ public class ProfileService(
     public async Task<ServiceResponse<NoContent>> Delete(Guid id)
     {
         var profile = await context.Profiles.FirstOrDefaultAsync(x => x.Id == id);
+        if (profile == null)
+            throw new KeyNotFoundException("Profile not found");
         context.Profiles.Remove(profile);
         await context.SaveChangesAsync();
-        
         var cacheKey = $"profile_{id}";
         await _cache.KeyDeleteAsync(cacheKey);
-
         return ServiceResponse<NoContent>.Success(StatusCodes.Status200OK);
     }
     
@@ -78,7 +75,7 @@ public class ProfileService(
         var dto = mapper.Map<ProfileDto>(profile);
         var response = await client.GetAsync($"{_mediaApiUrl}/{profile.Id}/3");
         await using var stream = await response.Content.ReadAsStreamAsync();
-        dto.ProfileImageUrl = (await JsonSerializer.DeserializeAsync<List<string>>(stream) ?? []).First();
+        dto.ProfileImageUrl = (await JsonSerializer.DeserializeAsync<ServiceResponse<List<string>>>(stream))?.Data!.First();
         return dto;
     }
 }
